+++ demos/browser/server.js
@@ -32,7 +32,7 @@ const log = message => {
   console.log(`${new Date().toISOString()} ${message}`);
 };
 
-const app = process.env.npm_config_app || 'meeting';
+const app = process.env.npm_config_app || 'meetingV2';
 
 const server = require(protocol).createServer(options, async (request, response) => {
   log(`${request.method} ${request.url} BEGIN`);